apps/oncolife/apps/patient-platform/patient-api/requirements.txt
# =============================================================================
# OncoLife Patient API - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework
# -----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# -----------------------------------------------------------------------------
# Configuration & Validation
# -----------------------------------------------------------------------------
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0

# -----------------------------------------------------------------------------
# AWS SDK
# -----------------------------------------------------------------------------
boto3>=1.34.0
botocore>=1.34.0

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
python-jose[cryptography]>=3.3.0

# -----------------------------------------------------------------------------
# Rate Limiting
# -----------------------------------------------------------------------------
slowapi>=0.1.9

# -----------------------------------------------------------------------------
# System Monitoring
# -----------------------------------------------------------------------------
psutil>=5.9.0

# -----------------------------------------------------------------------------
# HTTP Client
# -----------------------------------------------------------------------------
requests>=2.31.0
httpx>=0.25.0

# -----------------------------------------------------------------------------
# Date/Time
# -----------------------------------------------------------------------------
pytz>=2024.1

# -----------------------------------------------------------------------------
# Development & Testing (optional)
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
